Include target in makefile to git pull all repos (#34)
* Makefile now attempts to pull repos.
Previously a repo would be skipped if had already been checked out.  Now it does a pull so that a 'make' will bring
all of the repos up to date
* Update changelog
* Moved pulling functionality into it's own section
Having it git pull on every build can cause issues of speed and can cause conflicts.  Have a seperate
target that only updates all the repos is a fair compromise